Big changes badum badum- extracted the elasticsearch interfaces into a wrapper and a query builder
