+++ deploy.sh
@@ -163,7 +163,7 @@ DOCKER_COMPOSE_FILE="docker-compose.yml"
 
 export ESOPTS='{"timeout": 60, "max_retries": 3}' # 'timeout' parameter is deprecated
 export TERMFIELDS="article_title,text_content"
-export TERMAGGRS="top,significant,rare"
+export TERMAGGRS="top"
 export ELASTICSEARCH_INDEX_NAME_PREFIX="mc_search"
 export API_PORT
 export API_REPLICAS
