+++ test/__init__.py
@@ -1,6 +1,6 @@
 import os
 
-INDEX_NAME = "mediacloud_test"
+INDEX_NAME = "mc_search"
 ELASTICSEARCH_URL = "http://localhost:9200"
 FIXTURES_DIR = os.path.join(os.path.dirname(__file__), "fixtures")
 NUMBER_OF_TEST_STORIES = 5103
